requirements.txt
aniso8601==8.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Caching==1.9.0
Flask-RESTful==0.3.8
idna==2.9
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
MarkupSafe==1.1.1
marshmallow==3.6.1
numpy==1.19.0
pandas==1.0.5
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
simplejson==3.17.0
six==1.15.0
urllib3==1.25.9
webargs==6.1.0
Werkzeug==1.0.1
zipp==3.1.0
gunicorn==20.0.4
sentry-sdk[flask]==0.16.0